kmclifetime.cc dervies from kmccalculator now
