+++ src/libxtp/calculators/kmclifetime.h
@@ -19,27 +19,11 @@
 #ifndef __VOTCA_KMC_LIFETIME_H
 #define	__VOTCA_KMC_LIFETIME_H
 
-// #include <votca/xtp/vssmgroup.h>
-#include <vector>
-#include <map>
-#include <iostream>
-#include <fstream>
-#include <string>
-#include <stdio.h>
-#include <stdlib.h>
-#include <time.h>
-#include <cmath> // needed for abs(double)
-
-#include <votca/tools/vec.h>
-#include <votca/tools/matrix.h>
-#include <votca/tools/tokenizer.h>
-#include <votca/tools/globals.h>
-#include <votca/tools/random2.h>
 
 
-#include <votca/xtp/gnode.h>
-#include <math.h> // needed for fmod()
-#include <votca/ctp/qmcalculator.h>
+
+
+#include <votca/xtp/kmc.h>
 using namespace std;
 
 namespace votca { namespace xtp {
@@ -47,7 +31,7 @@ namespace votca { namespace xtp {
    
 
 
-class KMCLifetime : public ctp::QMCalculator 
+class KMCLifetime : public KMCCalculator 
 {
 public:
     KMCLifetime() {};
@@ -58,66 +42,27 @@ public:
 
 private:
        
-        
-            class Chargecarrier
-            {
-                public:
-                    Chargecarrier(): lifetime(0.0),steps(0) 
-                    {
-                        dr_travelled=tools::vec(0.0,0.0,0.0);
-                    }
-                    void updateLifetime(double dt) { lifetime+=dt;}
-                    void updateSteps(unsigned t) { steps+=t;}
-                    void resetCarrier() { lifetime=0;steps=0; dr_travelled=tools::vec(0.0,0.0,0.0);}
-                    const double& getLifetime(){return lifetime;}
-                    const unsigned& getSteps(){return steps;}
-                    
-                    int id;
-                    GNode *node;
-                    tools::vec dr_travelled;
-                    
-                private:
-                    double lifetime;
-                    unsigned steps;
-            };
+       
             
-            std::string CarrierInttoLongString(int carriertype);
-            std::string CarrierInttoShortString(int carriertype);
-            int StringtoCarriertype(std::string name);
-            
-	    void LoadGraph(ctp::Topology *top);
+	    
             void  RunVSSM(ctp::Topology *top);
-            void InitialRates();
-            void progressbar(double fraction);
+            
+            
             void ReadLifetimeFile( string filename);
-            double Promotetime(double cumulated_rate);
-            void ResetForbiddenlist(std::vector<int> &forbiddenid);
-            void AddtoForbiddenlist(int id, std::vector<int> &forbiddenid);
-            bool CheckForbidden(int id,const std::vector<int> &forbiddenlist);
-            bool CheckSurrounded(GNode* node,const std::vector<int> &forbiddendests);
-            GLink* ChooseHoppingDest(GNode* node);
             
             //tools::vec _field;
             
-            std::vector<GNode*> _nodes;
-            tools::Random2 * _RandomVariable;
             bool _do_carrierenergy;
             string _energy_outputfile;
             double _alpha;
-            string _injection_name;
-            string _injectionmethod;
             unsigned _outputsteps;
             unsigned int _insertions;
             std::string _lifetimefile;
             double _maxrealtime;
-            int _seed;
-            int _numberofcharges;
             string _trajectoryfile;
             int _carriertype;
-            double _temperature;
             string _outputfile;
             string _filename;
-            string _rates;
 };
